refactor(cactus-web): put Fieldset in separate file that both RadioGroup & CheckBoxGroup can use
CACTUS-215
