+++ modules/cactus-web/src/RadioGroup/RadioGroup.tsx
@@ -1,14 +1,12 @@
 import PropTypes from 'prop-types'
 import React from 'react'
-import styled from 'styled-components'
-import { margin, MarginProps, width, WidthProps } from 'styled-system'
+import { MarginProps, WidthProps } from 'styled-system'
 
 import { FieldProps, useAccessibleField } from '../AccessibleField/AccessibleField'
 import Box from '../Box/Box'
-import FieldWrapper from '../FieldWrapper/FieldWrapper'
+import Fieldset from '../Fieldset/Fieldset'
 import handleEvent from '../helpers/eventHandler'
 import { cloneAll } from '../helpers/react'
-import { border } from '../helpers/theme'
 import Label from '../Label/Label'
 import RadioButtonField, { RadioButtonFieldProps } from '../RadioButtonField/RadioButtonField'
 import StatusMessage from '../StatusMessage/StatusMessage'
@@ -207,37 +205,3 @@ const DefaultRadioGroup = RadioGroup as any
 DefaultRadioGroup.Button = RadioGroupButton
 
 export default DefaultRadioGroup as RadioGroupType
-
-// The `legend` is floated to get it to position the border correctly.
-const Fieldset = styled(FieldWrapper)<{ disabled?: boolean }>`
-  position: relative;
-  border: 0;
-  margin: 0;
-  padding: 0;
-  ${margin}
-  ${width}
-
-  legend {
-    box-sizing: border-box;
-    border-bottom: ${(p) => border(p.theme, 'currentcolor')};
-    padding-left: 16px;
-    padding-right: 28px;
-    width: 100%;
-    float: left;
-    + * {
-      clear: both;
-    }
-    color: ${(p) => (p.disabled ? p.theme.colors.mediumGray : 'currentcolor')};
-  }
-
-  ${Tooltip} {
-    position: absolute;
-    right: 8px;
-    top: 2px;
-    font-size: 16px;
-  }
-
-  ${StatusMessage} {
-    margin-top: 4px;
-  }
-`.withComponent('fieldset')
